renamed penalization to evidence
