+++ action/evidence/allegation.go
@@ -1,4 +1,4 @@
-package penalization
+package evidence
 
 import (
 	"encoding/binary"
