+++ action/evidence/allegation_test.go
@@ -1,4 +1,4 @@
-package penalization
+package evidence
 
 import (
 	"errors"
