+++ action/evidence/release.go
@@ -1,4 +1,4 @@
-package penalization
+package evidence
 
 import (
 	"encoding/json"
