+++ action/evidence/release_test.go
@@ -1,4 +1,4 @@
-package penalization
+package evidence
 
 import (
 	"testing"
